login

+++ API/Constants/BookingStatus.cs
@@ -2,9 +2,9 @@
 {
     public class BookingStatus
     {
-        public static readonly int Pending = 1;
-        public static readonly int Approved = 2;
-        public static readonly int Rejected = 3;
+        public static readonly decimal Pending = 1;
+        public static readonly decimal Approved = 2;
+        public static readonly decimal Rejected = 3;
         // should modified based on the requirement
         // admin vao co mot casi list danh sach nhung nguoi dat phong, co the approve hoac reject
     }
