+++ API/Constants/SystemRole.cs
@@ -7,10 +7,10 @@ namespace BookingApp.Constants
 {
     public static class SystemRole
     {
-        public static readonly int Admin = 1;
-        public static readonly int User = 2;
-        public static readonly int Operator = 3;
-        public static readonly int Guest = 4; 
+        public static readonly decimal Admin = 1;
+        public static readonly decimal User = 2;
+        public static readonly decimal Operator = 3;
+        public static readonly decimal Guest = 4; 
 
     }
 }
